requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# Environment Management
python-dotenv

# OpenAI Integration
openai

# Task Queue
celery
redis

# Utilities
pydantic

# Testing
pytest
pytest-asyncio
httpx
groq